Check if Feature is under click point and unique
features_at_point seems to use the bounding box and/or a buffer for each
feature. As a result the numbr of features returned is much smaller than
the whole set but not exactly what I want.  As such, lets loop through
the filtered features and check if the click point is really inside it

Also, my block-group shp file seems to have duplicate shapes in it.  I
need to clean that up; for now, the same layer shouldn't so we can just
filter.